+++ world-loading.js
@@ -3,6 +3,9 @@ import { MorphicEnv } from "./index.js";
 import { loadWorldFromResource } from "./serialization.js";
 import { resource, registerExtension as registerResourceExension } from "lively.resources";
 import { MorphicDB } from "./morphicdb/index.js";
+import { Path, date, promise } from "lively.lang";
+import { loadObjectFromPartsbinFolder } from "./partsbin.js";
+import LoadingIndicator from "./components/loading-indicator.js";
 
 
 export function pathForBrowserHistory(worldName, queryString) {
@@ -25,7 +28,7 @@ export async function loadWorldFromURL(url, oldWorld, options) {
 export async function loadWorldFromCommit(commitOrId, oldWorld, options) {
     let db = MorphicDB.default,
         newWorld = await db.load("world", undefined, options, commitOrId),
-        queryString = typeof document !== "undefined" ? document.location.search : ""  
+        queryString = typeof document !== "undefined" ? document.location.search : ""
   options = {
     pathForBrowserHistory: pathForBrowserHistory(newWorld.name, queryString),
     ...options
@@ -36,7 +39,7 @@ export async function loadWorldFromCommit(commitOrId, oldWorld, options) {
 export async function loadWorldFromDB(name, ref, oldWorld, options) {
   let db = MorphicDB.default,
       newWorld = await db.load("world", name, options, undefined/*commit||id*/, ref || undefined),
-      queryString = typeof document !== "undefined" ? document.location.search : ""  
+      queryString = typeof document !== "undefined" ? document.location.search : ""
   options = {
     pathForBrowserHistory: pathForBrowserHistory(name, queryString),
     ...options
@@ -132,4 +135,82 @@ async function setupLively2Lively() {
       client = await L2LClient.forLivelyInBrowser();
   console.log(`[lively] lively2lively client created ${client}`)
   return client;
+}
+
+
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+export async function interactivelySaveWorld(world, options) {
+  options = {showSaveDialog: true, ...options};
+
+  let name = world.name, tags = [], description = "",
+      oldCommit = Path("metadata.commit").get(world),
+      db = MorphicDB.default;
+
+  if (options.showSaveDialog) {
+    let dialog = await loadObjectFromPartsbinFolder("save world dialog");
+    ({name, tags, description} = await world.openPrompt(dialog, {targetWorld: world}));
+    if (!name) return null;
+  } else if (oldCommit) {
+    ({name, tags, description} = oldCommit);
+  }
+
+
+  var i = LoadingIndicator.open(`saving ${name}...`);
+  await promise.delay(80);
+
+  try {
+    let snapshotOptions = {previewWidth: 200, previewHeight: 200, previewType: "png"},
+        ref = "HEAD",
+        oldName = oldCommit ? oldCommit.name : world.name,
+        expectedParentCommit;
+
+    if (oldName !== name) {
+      let {exists, commitId: existingCommitId} = await db.exists("world", name);
+      if (exists) {
+        let overwrite = await world.confirm(`A world "${name}" already exists, overwrite?`);
+        if (!overwrite) return null;
+        expectedParentCommit = existingCommitId;
+      }
+      world.name = name;
+    } else {
+      expectedParentCommit = oldCommit ? oldCommit._id : undefined;
+    }
+
+    let commitSpec = {author: world.getCurrentUser(),
+                      message: "world save", tags, description},
+        commit = await db.snapshotAndCommit(
+          "world", name, world, snapshotOptions,
+          commitSpec, ref, expectedParentCommit);
+
+    // hist
+    if (window.history) {
+      let queryString = typeof document !== "undefined" ? document.location.search : "",
+          path = pathForBrowserHistory(name, queryString);
+      window.history.pushState({}, "lively.next", path);
+    }
+
+    world.setStatusMessage(`saved world ${name}`);
+    world.get("world-list") && world.get("world-list").onWorldSaved(name);
+
+    return commit;
+
+  } catch (err) {
+    let [_, typeAndName, expectedVersion, actualVersion] = err.message.match(/Trying to store "([^\"]+)" on top of expected version ([^\s]+) but ref HEAD is of version ([^\s\!]+)/) || [];
+    if (expectedVersion && actualVersion) {
+      let [newerCommit] = await db.log(actualVersion, 1, /*includeCommits = */true),
+          {author: {name: authorName}, timestamp} = newerCommit,
+          overwriteQ = `The current version of world ${name} is not the most recent!\n`
+      + `A newer version by ${authorName} was saved on `
+      + `${date.format(new Date(timestamp), "yyyy-mm-dd HH:MM")}. Overwrite?`,
+          overwrite = await world.confirm(overwriteQ);
+      if (!overwrite) return null;
+      world.changeMetaData("commit", newerCommit, /*serialize = */false, /*merge = */false);
+      return interactivelySaveWorld(world, {...options, showSaveDialog: false});
+    }
+
+    console.error(err);
+    world.logError("Error saving world: " + err);
+  } finally { i.remove(); }
+
 }
